requirements.txt
deeplite_torch_zoo==2.0.1
jinja2==3.1.2
tensorboard==2.14.0
torchvision==0.9.1

